+++ jest.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  testEnvironment: 'node',
+  testMatch: ['<rootDir>/test/**/*.test.ts'],
+  transform: {
+    '^.+\\.ts$': 'babel-jest',
+  },
+  moduleFileExtensions: ['js', 'ts', 'json'],
+  transformIgnorePatterns: ['/node_modules/'],
+};